[fix] 리뷰 작성자만 수정, 삭제할 수 있도록 수정

+++ src/main/java/com/prgrms/ohouse/domain/commerce/application/impl/ReviewServiceImpl.java
@@ -2,6 +2,7 @@ package com.prgrms.ohouse.domain.commerce.application.impl;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -22,6 +23,7 @@ import com.prgrms.ohouse.domain.commerce.model.review.exception.ReviewRegisterFa
 import com.prgrms.ohouse.domain.commerce.model.review.exception.ReviewUpdateFailException;
 import com.prgrms.ohouse.domain.common.file.FileIOException;
 import com.prgrms.ohouse.domain.common.file.FileManager;
+import com.prgrms.ohouse.domain.common.security.AuthUtility;
 import com.prgrms.ohouse.domain.user.model.User;
 import com.prgrms.ohouse.domain.user.model.UserRepository;
 
@@ -38,8 +40,9 @@ public class ReviewServiceImpl implements ReviewService {
 	private final ReviewRepository reviewRepository;
 	private final UserRepository userRepository;
 	private final FileManager fileManager;
+	private final AuthUtility authUtility;
 
-	//TODO: 로그인 된 사용자만 이용할 수 있도록 검증
+	//TODO: 로그인 된 사용자만 이용할 수 있도록 검증. Security 설정 변경?
 	@Transactional
 	@Override
 	public Long registerReview(ReviewRegisterCommand command) {
@@ -77,12 +80,13 @@ public class ReviewServiceImpl implements ReviewService {
 		return PagedPhotoReviewInformation.of(pageInformation, reviewPage.getContent());
 	}
 
-	//TODO: 리뷰 작성자만 지울 수 있도록 검증
 	@Transactional
 	@Override
 	public void deleteReview(Long id) {
 		Review review = reviewRepository.findById(id)
 			.orElseThrow(() -> new ReviewDeleteFailException("invalid review id"));
+		User authUser = authUtility.getAuthUser();
+		checkAuthor(authUser, review.getUser());
 		reviewRepository.delete(review);
 	}
 
@@ -90,9 +94,11 @@ public class ReviewServiceImpl implements ReviewService {
 	public void updateReview(ReviewUpdateCommand command) {
 		Review review = reviewRepository.findById(command.getId())
 			.orElseThrow(() -> new ReviewUpdateFailException("invalid review id"));
+		User authUser = authUtility.getAuthUser();
+		checkAuthor(authUser, review.getUser());
 		if (command.isPhotoReview()) {
 			try {
-				fileManager.delete(review.getReviewImage());
+				fileManager.delete(review.getReviewImage(), review);
 				fileManager.store(command.getReviewImage(), review);
 			} catch (FileIOException e) {
 				throw new ReviewUpdateFailException(e.getMessage(), e);
@@ -100,4 +106,10 @@ public class ReviewServiceImpl implements ReviewService {
 		}
 		review.modifyReview(command.getReviewPoint(), command.getContents());
 	}
+
+	private void checkAuthor(User target, User author) {
+		if (target != author)
+			throw new AccessDeniedException("access denied");
+	}
+
 }
